Ignore some pathological cases around tag values
Hypothesis exposed a weird edge case where you have `@a(\) @a(0)` at the
end of an item.  Although it's possible to fix this by rearranging parts
of the regex, I've accepted that this is just weird and gone for
consistency with TaskPaper.app.

See also: http://support.hogbaysoftware.com/t/bug-pathological-edge-case-for-the-tag-matching-regex/1815